Extract inner classes to make their relations explicit and work easier with the code
